pyproject.toml
[project]
name = "beguest-helper-room-info"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0",
    "gspread>=6.2.0",
    "loguru>=0.7.3",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "yadisk>=3.2.0",
]
